Move driver testing into its own package and add additional test helpers
